Moving reports and picklist and print reducers
