+++ src/components/AppComponent.js
@@ -5,7 +5,7 @@ import { IndexLink, hashHistory } from 'react-router';
 import { LinkContainer } from 'react-router-bootstrap';
 import { Navbar, Nav, NavItem } from 'react-bootstrap';
 import FontAwesome from 'react-fontawesome';
-import { TYPES as PICK_TYPES } from '../reducers/picklist';
+import { TYPES as PICK_TYPES } from '../modules/picklist/picklistReducer';
 import MusitUserAccount from './UserAccount';
 import './AppComponent.css';
 import Logo from './musitLogo.png';
