+++ src/modules/reports/__tests__/reportsReducer-test.js
@@ -1,8 +1,8 @@
 import configureMockStore from 'redux-mock-store';
 import createMiddleware from '../../../redux/clientMiddleware';
 import ApiClient from '../../../redux/ApiClient';
-import * as actions from '../index';
-import reducer from '../index';
+import * as actions from '../reportsReducer';
+import reducer from '../reportsReducer';
 import Config from '../../../config';
 import request from 'superagent';
 import nocker from 'superagent-nock';
